requirements.txt
tqdm==4.65.0
torchvision==0.15.2
torchaudio==2.0.2
torch==2.0.1
xformers==0.0.20
timm==0.9.12
scikit-learn-extra==0.3.0
scikit-learn==1.3.0
scikit-image==0.21
pyyaml==6.0.1
pillow==9.3.0
opencv-python==4.8.0.74
numpy==1.24.1
matplotlib==3.7.2
loguru==0.7.2
h5py==3.10.0
einops==0.7.0
scipy==1.11.1
distinctipy==1.3.4